# openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BhashaAI Backend",
    "description": "\n## BhashaAI - Generative AI for Education & Gujarati Language Technologies\n\nA comprehensive API for educational AI services supporting:\n\n- **Question Paper Generation** - AI-powered question paper creation from educational content\n- **Assignment Help** - Socratic tutoring and step-by-step solutions\n- **Teaching Tools** - Mind maps, slides, and study materials generation\n- **Paper Checking** - OCR-based automated evaluation with Gujarati support\n- **Gujarati Learning** - Vocabulary, grammar, and gamified learning\n- **Audio Services** - STT/TTS for Gujarati language\n\n### Authentication\nAll protected endpoints require JWT Bearer token authentication.\n\n### Bilingual Support\nResponses include both English and Gujarati (ગુજરાતી) messages.\n        ",
    "contact": {
      "name": "BhashaAI Team",
      "email": "support@bhashaai.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/worksheets/generate": {
      "post": {
        "tags": [
          "Worksheets"
        ],
        "summary": "Generate Worksheet",
        "description": "Generate a new step-by-step worksheet using AI.",
        "operationId": "generate_worksheet_api_v1_worksheets_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorksheetGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorksheetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/worksheets/": {
      "get": {
        "tags": [
          "Worksheets"
        ],
        "summary": "List Worksheets",
        "description": "List user's worksheets.",
        "operationId": "list_worksheets_api_v1_worksheets__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorksheetResponse"
                  },
                  "type": "array",
                  "title": "Response List Worksheets Api V1 Worksheets  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/worksheets/{worksheet_id}": {
      "get": {
        "tags": [
          "Worksheets"
        ],
        "summary": "Get Worksheet",
        "description": "Get worksheet details.",
        "operationId": "get_worksheet_api_v1_worksheets__worksheet_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worksheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Worksheet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorksheetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/worksheets/{worksheet_id}/attempts": {
      "post": {
        "tags": [
          "Worksheets"
        ],
        "summary": "Start Attempt",
        "description": "Start a new attempt (game session) for a worksheet.",
        "operationId": "start_attempt_api_v1_worksheets__worksheet_id__attempts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "worksheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Worksheet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/worksheets/attempts/{attempt_id}": {
      "get": {
        "tags": [
          "Worksheets"
        ],
        "summary": "Get Attempt",
        "description": "Get attempt state.",
        "operationId": "get_attempt_api_v1_worksheets_attempts__attempt_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/worksheets/attempts/{attempt_id}/step": {
      "post": {
        "tags": [
          "Worksheets"
        ],
        "summary": "Submit Step",
        "description": "Submit an answer for a single step.\nReturns feedback and next step info.",
        "operationId": "submit_step_api_v1_worksheets_attempts__attempt_id__step_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitStepRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepFeedback"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AttemptResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "worksheet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Worksheet Id"
          },
          "current_question_index": {
            "type": "integer",
            "title": "Current Question Index"
          },
          "current_step_index": {
            "type": "integer",
            "title": "Current Step Index"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Progress Data"
          }
        },
        "type": "object",
        "required": [
          "id",
          "worksheet_id",
          "current_question_index",
          "current_step_index",
          "score",
          "status",
          "progress_data"
        ],
        "title": "AttemptResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StepFeedback": {
        "properties": {
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "points_awarded": {
            "type": "integer",
            "title": "Points Awarded"
          },
          "next_step_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Step Index"
          },
          "next_question_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Question Index"
          },
          "is_complete": {
            "type": "boolean",
            "title": "Is Complete",
            "default": false
          },
          "correct_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Answer"
          }
        },
        "type": "object",
        "required": [
          "is_correct",
          "message",
          "points_awarded"
        ],
        "title": "StepFeedback"
      },
      "SubmitStepRequest": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "step_answer": {
            "type": "string",
            "title": "Step Answer"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "step_answer"
        ],
        "title": "SubmitStepRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorksheetGenerateRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "grade_level": {
            "type": "string",
            "title": "Grade Level"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "medium"
          },
          "num_questions": {
            "type": "integer",
            "title": "Num Questions",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "topic",
          "subject",
          "grade_level"
        ],
        "title": "WorksheetGenerateRequest"
      },
      "WorksheetQuestionBase": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorksheetStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          }
        },
        "type": "object",
        "required": [
          "content",
          "steps",
          "correct_answer",
          "order"
        ],
        "title": "WorksheetQuestionBase"
      },
      "WorksheetResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "grade_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Level"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "medium"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/WorksheetQuestionBase"
            },
            "type": "array",
            "title": "Questions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "topic",
          "subject",
          "id",
          "user_id",
          "status",
          "created_at"
        ],
        "title": "WorksheetResponse"
      },
      "WorksheetStep": {
        "properties": {
          "step_text": {
            "type": "string",
            "title": "Step Text"
          },
          "answer_key": {
            "type": "string",
            "title": "Answer Key"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          }
        },
        "type": "object",
        "required": [
          "step_text",
          "answer_key"
        ],
        "title": "WorksheetStep"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /api/v1/auth/login"
      },
      "RefreshToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Refresh-Token",
        "description": "JWT refresh token for obtaining new access tokens"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.bhashaai.com",
      "description": "Production server"
    }
  ]
}